+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.3]
+
 ## Fixed
 
 - Fixed up `GitCommitRef` `change_counts` field type
 - Added back some required `build` structure Option wrappers
 - Fixed formatting of date-time URL parameters
+- Added `ims` (Identity Management) example: `ims_query`
+- Fixed Pull Request create function `git::pull_requests::create()`
+- Added `git_pr_create` example
 
 ## [0.5.2]
 
@@ -123,7 +128,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.2...HEAD
+[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.3...HEAD
+[0.5.3]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.2...0.5.3
 [0.5.2]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/microsoft/azure-devops-rust-api/compare/0.4.0...0.5.0